// src/app.enum.ts
export enum WS_CMD {
  MOUSE_UP = 'mouse_up',
  MOUSE_DOWN = 'mouse_down',
  MOUSE_LEFT = 'mouse_left',
  MOUSE_RIGHT = 'mouse_right',
  MOUSE_POSITION = 'mouse_position',
  DRAW_CIRCLE = 'draw_circle',
  DRAW_RECTANGLE = 'draw_rectangle',
  DRAW_SQUARE = 'draw_square',
  PRNT_SCRN = 'prnt_scrn',
}